[live share]: refactor
