+++ src/extension.ts
@@ -25,19 +25,19 @@ import {
 } from './configuration';
 import { changeColor } from './color-library';
 import { Logger } from './logging';
-import { addVSLSIntegration } from './vsls/integration';
+import { addLiveShareIntegration } from './live-share/integration';
 
 const { commands, workspace } = vscode;
 
 export async function activate(context: vscode.ExtensionContext) {
   console.log('Extension "vscode-peacock" is now active!');
 
-  addVSLSIntegration(context);
-
   registerCommands();
   addSubscriptions(context);
   await initializeTheStarterSetOfFavorites(context);
   await applyInitialConfiguration();
+
+  addLiveShareIntegration(context);
 }
 
 function addSubscriptions(context: vscode.ExtensionContext) {
