+++ src/live-share/constants.ts
@@ -1,0 +1,3 @@
+export const VSLS_SHARE_COLOR_MEMENTO_NAME = 'peacock.vsls.share.color';
+export const VSLS_JOIN_COLOR_MEMENTO_NAME = 'peacock.vsls.join.color';
+export const VSCODE_WORKSPACE_COLORS_SETTING_NAME = 'workbench.colorCustomizations';
