+++ src/live-share/integration.ts
@@ -1,51 +1,39 @@
 import * as vsls from 'vsls';
 import * as vscode from 'vscode';
 
-import { VSLS_SHARE_COLOR_MEMENTO_NAME, VSLS_JOIN_COLOR_MEMENTO_NAME } from './constants';
+import { VSLS_SHARE_COLOR_MEMENTO_NAME, VSLS_JOIN_COLOR_MEMENTO_NAME, VSCODE_WORKSPACE_COLORS_SETTING_NAME } from './constants';
 import { changeColor } from '../color-library';
 import { registerLiveShareIntegrationCommands } from './liveshare-commands';
 import { extensionContext, setExtensionContext } from './extensionContext';
 
-let priorWorkspaceColorCustomizations: any;
-export const revertLiveShareWorkspaceColors = async () => {
+let peacockColorCustomizations: any;
+export async function revertLiveShareWorkspaceColors() {
     await vscode.workspace
         .getConfiguration()
         .update(
-            'workbench.colorCustomizations',
-            priorWorkspaceColorCustomizations,
+            VSCODE_WORKSPACE_COLORS_SETTING_NAME,
+            peacockColorCustomizations,
             vscode.ConfigurationTarget.Workspace
         );
 
-    priorWorkspaceColorCustomizations = null;
+    peacockColorCustomizations = null;
 };
 
-const setLiveShareSessionWorkspaceColors = async (isHost: boolean) => {
+async function setLiveShareSessionWorkspaceColors(isHost: boolean) {
     const colorSettingName = (isHost)
         ? VSLS_SHARE_COLOR_MEMENTO_NAME
         : VSLS_JOIN_COLOR_MEMENTO_NAME;
 
     const liveShareColorSetting = await extensionContext.globalState.get<string>(colorSettingName);
-
     if (!liveShareColorSetting) {
         return;
-    }
-
-    // const vslsApi = (await vsls.getApi())!;
-    // const isLiveShareSession = (vslsApi && vslsApi.session.id);
-
-    // we need to update `priorWorkspaceColorCustomizations` only when it is `undefined`
-    // to prevent the case of multiple color changes during live share session
-    if (!priorWorkspaceColorCustomizations) {
-        priorWorkspaceColorCustomizations = await vscode.workspace
-            .getConfiguration()
-            .get('workbench.colorCustomizations');
     }
 
     await changeColor(liveShareColorSetting);
 }
 
-export const refreshLiveShareSessionColor = async (): Promise<boolean> => {
-    const vslsApi = (await vsls.getApi())!;
+export async function refreshLiveShareSessionColor(): Promise<boolean> {
+    const vslsApi = (await vsls.getApi());
 
     if (!vslsApi || !vslsApi.session.id) {
         return false;
@@ -56,24 +44,30 @@ export const refreshLiveShareSessionColor = async (): Promise<boolean> => {
     return true;
 }
 
-export const addVSLSIntegration = async (context: vscode.ExtensionContext) => {
+export async function addLiveShareIntegration(context: vscode.ExtensionContext) {
     setExtensionContext(context);
 
     registerLiveShareIntegrationCommands();
     
-    const vslsApi = (await vsls.getApi())!;
+    const vslsApi = (await vsls.getApi());
     await vscode.commands.executeCommand('setContext', 'peacock:liveshare', !!vslsApi);
 
     if (!vslsApi) {
         return;
     }
 
-    vslsApi!.onDidChangeSession(async (e) => {
+    vslsApi!.onDidChangeSession(async function onLiveShareSessionCHange(e) {
         // If there isn't a session ID, then that
         // means the session has been ended.
         if (!e.session.id) {
           return await revertLiveShareWorkspaceColors();
         }
+
+        // we need to update `peacockColorCustomizations` only when it is `undefined`
+        // to prevent the case of multiple color changes during live share session
+        peacockColorCustomizations = await vscode.workspace
+            .getConfiguration()
+            .get(VSCODE_WORKSPACE_COLORS_SETTING_NAME);
     
         const isHost = (e.session.role === vsls.Role.Host);
         return await setLiveShareSessionWorkspaceColors(isHost);
