--- src/vsls/constants.ts
@@ -1,2 +1,0 @@
-export const VSLS_SHARE_COLOR_MEMENTO_NAME = 'peacock.vsls.share.color';
-export const VSLS_JOIN_COLOR_MEMENTO_NAME = 'peacock.vsls.join.color';
